requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery[redis]
redis
requests
python-dotenv
apscheduler
